backend/requirements.txt
flask>=2.3.0
flask-cors==3.0.10
boto3==1.28.38
opensearch-py==2.0.0
mem0ai>=0.1.55
python-dotenv==0.19.1
